Add file system directory enumeration function
There are a lot of different formats that can be used to iterate
directories based on the system call being implemented.  Rather than
force them to convert something arbitrary like linux_dirent64 (which has
its own problems since the size of the structure factors into the
iteration), use a caller-supplied callback instead so the system call
can do whatever it wants with the information.