Fix UAPI header generation such that long converts to __int3264
Without setting the proper target arguments to libclang, long was
defaulting to 4 bytes in the generated x64 headers, which is incorrect.
GCC/linux treat long as 8 bytes on the x64 platform.  Visual C++
supplies a special __int3264 data type that compiles to 4 bytes on
32-bit builds and 8 bytes on 64-bit builds, which solves the problem.
Additionally after setting the targets for clang the alignment for 64
bit POD types in 32-bit code changed.  Visual C++ uses 8 byte alignment
in this case, GCC uses 4 byte alignment.  It shouldn't be problematic
for POD types to be aligned differently so just suppress generation of
the __alignof() static_assert for POD types.